pyproject.toml
[tool.poetry]
name = "webtronic_fastapi"
version = "0.1.0"
description = ""
authors = ["RomanL91 <dev_roman91@mail.ru>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.89.0"
uvicorn = {extras = ["standart"], version = "^0.20.0"}
motor = "^3.1.1"
pydantic = {extras = ["email"], version = "^1.10.4"}
python-multipart = "^0.0.5"
passlib = "^1.7.4"
PyJWT = "^2.6.0"
envparse = "^0.2.0"
bcrypt = "^4.0.1"

[tool.poetry.dev-dependencies]
autopep8 = "^2.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
